+++ src/lib/components/ContractSelectSection.tsx
@@ -12,7 +12,10 @@ import type { ContractAddr, Option } from "lib/types";
 
 import { ExplorerLink } from "./ExplorerLink";
 import { EditContractDetails, SaveContractDetails } from "./modal";
-import { SelectContract } from "./modal/select-contract";
+import {
+  SelectContractAdmin,
+  SelectContractInstantiator,
+} from "./modal/select-contract";
 
 interface DisplayNameProps {
   notSelected: boolean;
@@ -29,6 +32,7 @@ interface ContractDetailsButtonProps {
 }
 
 interface ContractSelectSectionProps {
+  mode: "instantiator" | "admin";
   contractAddress: ContractAddr;
   onContractSelect: (contract: ContractAddr) => void;
 }
@@ -97,7 +101,7 @@ const ContractDetailsButton = ({
 };
 
 export const ContractSelectSection = observer(
-  ({ contractAddress, onContractSelect }: ContractSelectSectionProps) => {
+  ({ mode, contractAddress, onContractSelect }: ContractSelectSectionProps) => {
     const { getContractLocalInfo } = useContractStore();
     const isMobile = useMobile();
     const endpoint = useEndpoint();
@@ -160,9 +164,13 @@ export const ContractSelectSection = observer(
         fontSize="12px"
         justify="space-between"
         align="center"
+        width="full"
       >
-        <Flex gap="24px" width="70%">
-          <Flex direction="column" width="70%">
+        <Flex gap="24px" width={mode === "instantiator" ? "70%" : "60%"}>
+          <Flex
+            direction="column"
+            width={mode === "instantiator" ? "70%" : "35%"}
+          >
             Contract Address
             {!notSelected ? (
               <ExplorerLink
@@ -170,7 +178,9 @@ export const ContractSelectSection = observer(
                 type="contract_address"
                 canCopyWithHover
                 // TODO - Revisit not necessary if disable UI for mobile is implemented
-                textFormat={isMobile ? "truncate" : "normal"}
+                textFormat={
+                  isMobile || mode === "admin" ? "truncate" : "normal"
+                }
                 maxWidth="none"
               />
             ) : (
@@ -179,7 +189,12 @@ export const ContractSelectSection = observer(
               </Text>
             )}
           </Flex>
-          <Flex direction="column" width="calc(30% - 24px)">
+          <Flex
+            direction="column"
+            width={
+              mode === "instantiator" ? "calc(30% - 24px)" : "calc(65% - 24px)"
+            }
+          >
             Contract Name
             <DisplayName
               notSelected={notSelected}
@@ -191,7 +206,7 @@ export const ContractSelectSection = observer(
         </Flex>
 
         <Flex gap="8px">
-          {contractState.isValid && (
+          {mode === "instantiator" && contractState.isValid && (
             <ContractDetailsButton
               contractAddress={contractAddress}
               contractLocalInfo={contractLocalInfo}
@@ -199,10 +214,17 @@ export const ContractSelectSection = observer(
               label={contractState.label}
             />
           )}
-          <SelectContract
-            notSelected={notSelected}
-            onContractSelect={onContractSelect}
-          />
+          {mode === "instantiator" ? (
+            <SelectContractInstantiator
+              notSelected={notSelected}
+              onContractSelect={onContractSelect}
+            />
+          ) : (
+            <SelectContractAdmin
+              notSelected={notSelected}
+              onContractSelect={onContractSelect}
+            />
+          )}
         </Flex>
       </Flex>
     );