+++ src/lib/components/state/ZeroState.tsx
@@ -4,26 +4,24 @@ import { useRouter } from "next/router";
 import { MdOutlineAdd, MdBookmarkBorder, MdSearch } from "react-icons/md";
 
 import { SaveNewContract } from "lib/components/modal/contract";
+import { ADMIN_SPECIAL_SLUG, INSTANTIATED_LIST_NAME } from "lib/data";
 import type { LVPair } from "lib/types";
+import { formatSlugName } from "lib/utils";
 
 import { DisconnectedState } from "./DisconnectedState";
 
 interface ZeroStateProps {
   list: LVPair;
   isReadOnly?: boolean;
-  isInstantiatedByMe: boolean;
 }
 
-const ActionSection = ({
-  isInstantiatedByMe,
-  handleAction,
-  list,
-}: {
-  isInstantiatedByMe: boolean;
-  handleAction?: () => void;
+interface ActionSectionProps {
   list: LVPair;
-}) => {
-  return isInstantiatedByMe ? (
+  handleAction?: () => void;
+}
+
+const ActionSection = ({ list, handleAction }: ActionSectionProps) =>
+  list.value === formatSlugName(INSTANTIATED_LIST_NAME) ? (
     <Button rightIcon={<MdOutlineAdd />} onClick={handleAction}>
       Deploy New Contract
     </Button>
@@ -44,20 +42,29 @@ const ActionSection = ({
       Created contract list and saved contracts are stored in your device only.
     </Flex>
   );
-};
 
 /**
  *
  * @todo Will be refactored in the next PR
  */
 
-export const ZeroState = ({
-  list,
-  isReadOnly,
-  isInstantiatedByMe,
-}: ZeroStateProps) => {
+export const ZeroState = ({ list, isReadOnly }: ZeroStateProps) => {
   const router = useRouter();
   const { isWalletConnected } = useWallet();
+
+  const isInstantiatedByMe =
+    list.value === formatSlugName(INSTANTIATED_LIST_NAME);
+
+  const renderText = () => {
+    switch (list.value) {
+      case formatSlugName(INSTANTIATED_LIST_NAME):
+        return "Your deployed contract through this address will display here.";
+      case ADMIN_SPECIAL_SLUG:
+        return "You don’t have any admin access to any contracts.";
+      default:
+        return "You don’t have any saved contracts.";
+    }
+  };
 
   return (
     <Flex
@@ -73,14 +80,9 @@ export const ZeroState = ({
       ) : (
         <Flex alignItems="center" flexDir="column" gap="4">
           <Icon as={MdSearch} color="gray.600" boxSize="16" />
-          <Text color="text.dark">
-            {isInstantiatedByMe
-              ? "Your deployed contract through this address will display here"
-              : "You don’t have any saved contracts."}
-          </Text>
+          <Text color="text.dark">{renderText()}</Text>
           {!isReadOnly && (
             <ActionSection
-              isInstantiatedByMe={isInstantiatedByMe}
               list={list}
               handleAction={() => router.push("/deploy")}
             />