+++ src/lib/pages/execute/index.tsx
@@ -101,6 +101,7 @@ const Execute = () => {
       />
 
       <ContractSelectSection
+        mode="instantiator"
         contractAddress={watchContractAddress}
         onContractSelect={onContractSelect}
       />