+++ src/lib/pages/migrate/types.ts
@@ -1,0 +1,8 @@
+import type { ContractAddr, HumanAddr, Option } from "lib/types";
+
+export interface MigratePageState {
+  isValid: boolean;
+  contractAddress: ContractAddr;
+  admin: Option<ContractAddr | HumanAddr>;
+  codeId: Option<number>;
+}